backend/requirements.txt
boto3
flask
flask-cors
python-dotenv
pywin32